🗺️ Atlas: [Architectural Refactor] Split Conversion Blob
Extracted complex `src/semantic/conversion.rs` (~2880 lines) into highly-cohesive, loosely-coupled submodules in `src/semantic/conversion/` (e.g., `values.rs`, `bindings.rs`, `assertions.rs`). Handled visibility properly with `pub(crate)` and ensured all imports, tests, and dependencies remain flawlessly functioning. Updated `.jules/atlas.md` with architectural decision logs.
