+++ src/semantic/conversion/values.rs
@@ -1,0 +1,637 @@
+use crate::ast::Expr;
+use crate::errors::GlossaError;
+use crate::semantic::assembly::{AssembledStatement, Constituent, Literal};
+use crate::semantic::expressions::analyze_argument_expr;
+use crate::semantic::model::*;
+use crate::semantic::resolver::Scope;
+use crate::semantic::types::GlossaType;
+
+use super::*;
+
+/// Helper: Detect iterator pattern
+pub(crate) fn classify_iterator_pattern(
+    asm_stmt: &AssembledStatement,
+    scope: &mut Scope,
+) -> Result<Option<AnalyzedStatement>, GlossaError> {
+    let has_find_or_print_verb = if let Some(ref verb) = asm_stmt.verb {
+        let verb_lemma = &verb.lemma;
+        crate::morphology::lexicon::is_print_verb(verb_lemma)
+            || crate::morphology::lexicon::is_find_verb(verb_lemma)
+    } else {
+        false
+    };
+
+    if (!asm_stmt.participles.is_empty()
+        || !asm_stmt.adjectives.is_empty()
+        || has_find_or_print_verb)
+        && let Some(analyzed) = detect_iterator_pattern(asm_stmt, scope)?
+    {
+        return Ok(Some(AnalyzedStatement::Print(vec![analyzed])));
+    }
+
+    Ok(None)
+}
+
+/// Helper: Detect property access print pattern (pi.xi)
+pub(crate) fn classify_property_access_print(
+    asm_stmt: &AssembledStatement,
+    scope: &mut Scope,
+) -> Result<Option<AnalyzedStatement>, GlossaError> {
+    let Some(verb) = &asm_stmt.verb else {
+        return Ok(None);
+    };
+
+    if !crate::morphology::lexicon::is_print_verb(&verb.lemma) {
+        return Ok(None);
+    }
+
+    if asm_stmt.genitives.is_empty() {
+        return Ok(None);
+    }
+
+    let Some(subject) = &asm_stmt.subject else {
+        return Ok(None);
+    };
+
+    // Get owner from genitive (use lemma to get base variable name)
+    let owner_lemma = &asm_stmt.genitives[0].lemma;
+
+    // Get property from subject (nominative)
+    let property = &subject.normalized;
+
+    // Check if owner is a struct type in scope
+    let Some(owner_type) = scope.lookup(owner_lemma) else {
+        return Ok(None);
+    };
+
+    if !matches!(owner_type, GlossaType::Struct { .. }) {
+        return Ok(None);
+    }
+
+    // Build property access
+    let prop_access = AnalyzedExpr {
+        expr: AnalyzedExprKind::PropertyAccess {
+            owner: Box::new(AnalyzedExpr {
+                expr: AnalyzedExprKind::Variable(owner_lemma.clone()),
+                glossa_type: owner_type.clone(),
+            }),
+            property: property.clone(),
+        },
+        glossa_type: GlossaType::Unknown, // TODO: Look up field type
+    };
+
+    Ok(Some(AnalyzedStatement::Print(vec![prop_access])))
+}
+
+/// Helper: Detect subjunctive comparison (which looks like binding verb but isn't)
+pub(crate) fn classify_subjunctive_comparison(
+    asm_stmt: &AssembledStatement,
+    scope: &mut Scope,
+) -> Result<Option<AnalyzedStatement>, GlossaError> {
+    let Some(verb) = &asm_stmt.verb else {
+        return Ok(None);
+    };
+
+    if !crate::morphology::lexicon::is_binding_verb(&verb.lemma)
+        || asm_stmt.operators.is_empty()
+        || asm_stmt.literals.is_empty()
+        || verb.mood != Some(crate::morphology::Mood::Subjunctive)
+    {
+        return Ok(None);
+    }
+
+    let Some(subject) = &asm_stmt.subject else {
+        return Ok(None);
+    };
+
+    let left = if let Some(var_type) = scope.lookup(&subject.lemma) {
+        AnalyzedExpr {
+            expr: AnalyzedExprKind::Variable(subject.lemma.clone()),
+            glossa_type: var_type.clone(),
+        }
+    } else {
+        AnalyzedExpr {
+            expr: AnalyzedExprKind::BooleanLiteral(false),
+            glossa_type: GlossaType::Boolean,
+        }
+    };
+
+    let right = literal_to_analyzed_expr(&asm_stmt.literals[0]);
+    let op = asm_stmt.operators[0];
+    let comparison = build_binary_expr(left, op, right);
+
+    Ok(Some(AnalyzedStatement::Expression(vec![comparison])))
+}
+
+/// Helper: Default expression
+#[allow(unused_variables)]
+pub(crate) fn classify_expression(
+    asm_stmt: &AssembledStatement,
+    scope: &Scope,
+) -> Result<AnalyzedStatement, GlossaError> {
+    // Determine if we should attempt to build expressions from literals+operators
+    // or if we are in a fallback scenario (using Subject/Object with operators).
+    // If literals < operators + 1, build_expressions_from_literals_and_ops will fail.
+    // In that case, we only build literals and let the fallback logic handle operators.
+
+    let (literals_to_build, operators_to_build) = if !asm_stmt.operators.is_empty()
+        && asm_stmt.literals.len() < asm_stmt.operators.len() + 1
+    {
+        // Fallback case: operators likely depend on Subject/Object
+        (asm_stmt.literals.as_slice(), &[][..])
+    } else {
+        (asm_stmt.literals.as_slice(), asm_stmt.operators.as_slice())
+    };
+
+    let mut exprs = build_expressions_from_literals_and_ops(literals_to_build, operators_to_build)?;
+
+    // If we have operators but couldn't build a full expression from literals alone (usually implies literals < 2),
+    // we should look for Subject/Object to complete the binary expression.
+    if !asm_stmt.operators.is_empty() && asm_stmt.literals.len() < 2 {
+        let op = asm_stmt.operators[0];
+
+        let left = asm_stmt.subject.as_ref().map(|subj| AnalyzedExpr {
+            expr: AnalyzedExprKind::Variable(subj.lemma.clone()),
+            glossa_type: GlossaType::Unknown,
+        });
+
+        // Try to get right operand from exprs (literal) or object or nominatives
+        let right = if let Some(lit_expr) = exprs.first() {
+            Some(lit_expr.clone())
+        } else if let Some(ref obj) = asm_stmt.object {
+            Some(AnalyzedExpr {
+                expr: AnalyzedExprKind::Variable(obj.lemma.clone()),
+                glossa_type: GlossaType::Unknown,
+            })
+        } else {
+            asm_stmt.nominatives.first().map(|nom| AnalyzedExpr {
+                expr: AnalyzedExprKind::Variable(nom.lemma.clone()),
+                glossa_type: GlossaType::Unknown,
+            })
+        };
+
+        if let (Some(l), Some(r)) = (left, right) {
+            let bin_expr = build_binary_expr(l, op, r);
+            exprs = vec![bin_expr];
+        }
+    }
+
+    // Fallback: If no literals/ops, check Subject/Object
+    if exprs.is_empty() {
+        if let Some(ref subj) = asm_stmt.subject {
+            exprs.push(AnalyzedExpr {
+                expr: AnalyzedExprKind::Variable(subj.lemma.clone()),
+                glossa_type: GlossaType::Unknown,
+            });
+        } else if let Some(ref obj) = asm_stmt.object {
+            exprs.push(AnalyzedExpr {
+                expr: AnalyzedExprKind::Variable(obj.lemma.clone()),
+                glossa_type: GlossaType::Unknown,
+            });
+        }
+    }
+
+    if asm_stmt.is_propagate && !exprs.is_empty() {
+        #[allow(clippy::collapsible_if)]
+        if let Some(last_expr) = exprs.pop() {
+            let try_expr = AnalyzedExpr {
+                glossa_type: last_expr.glossa_type.clone(),
+                expr: AnalyzedExprKind::Try(Box::new(last_expr)),
+            };
+            exprs.push(try_expr);
+        }
+    }
+
+    Ok(AnalyzedStatement::Expression(exprs))
+}
+
+/// Helper: Detect Enum variant (None, Some, Ok, Err)
+pub(crate) fn detect_enum_variant(
+    word: &Constituent,
+    literals: &[Literal],
+) -> Option<(AnalyzedExpr, GlossaType)> {
+    let lemma = &word.lemma;
+    let original = &word.normalized;
+
+    // Helper to check if a word matches a predicate
+    let check = |pred: fn(&str) -> bool| pred(lemma) || pred(original);
+
+    // None
+    if check(crate::morphology::lexicon::is_none_word) {
+        return Some((
+            AnalyzedExpr {
+                expr: AnalyzedExprKind::None,
+                glossa_type: GlossaType::Option(Box::new(GlossaType::Unknown)),
+            },
+            GlossaType::Option(Box::new(GlossaType::Unknown)),
+        ));
+    }
+
+    if let Some(lit) = literals.first() {
+        let inner_expr = literal_to_analyzed_expr(lit);
+        let inner_type = inner_expr.glossa_type.clone();
+
+        // Some
+        if check(crate::morphology::lexicon::is_some_word) {
+            return Some((
+                AnalyzedExpr {
+                    expr: AnalyzedExprKind::Some(Box::new(inner_expr)),
+                    glossa_type: GlossaType::Option(Box::new(inner_type.clone())),
+                },
+                GlossaType::Option(Box::new(inner_type)),
+            ));
+        }
+
+        // Ok
+        if check(crate::morphology::lexicon::is_ok_word) {
+            return Some((
+                AnalyzedExpr {
+                    expr: AnalyzedExprKind::Ok(Box::new(inner_expr)),
+                    glossa_type: GlossaType::Result(
+                        Box::new(inner_type.clone()),
+                        Box::new(GlossaType::String),
+                    ),
+                },
+                GlossaType::Result(Box::new(inner_type), Box::new(GlossaType::String)),
+            ));
+        }
+
+        // Err
+        if check(crate::morphology::lexicon::is_err_word) {
+            return Some((
+                AnalyzedExpr {
+                    expr: AnalyzedExprKind::Err(Box::new(inner_expr)),
+                    glossa_type: GlossaType::Result(
+                        Box::new(GlossaType::Unknown),
+                        Box::new(inner_type.clone()),
+                    ),
+                },
+                GlossaType::Result(Box::new(GlossaType::Unknown), Box::new(inner_type)),
+            ));
+        }
+    }
+
+    None
+}
+
+pub(crate) fn extract_unwrap(
+    asm_stmt: &AssembledStatement,
+    scope: &Scope,
+) -> Result<Option<(AnalyzedExpr, GlossaType)>, GlossaError> {
+    if !asm_stmt.unwraps.is_empty() {
+        let inner_analyzed = analyze_argument_expr(&asm_stmt.unwraps[0], scope)?;
+        return Ok(Some((
+            AnalyzedExpr {
+                expr: AnalyzedExprKind::Unwrap(Box::new(inner_analyzed)),
+                glossa_type: GlossaType::Unknown, // Type will be inferred
+            },
+            GlossaType::Unknown,
+        )));
+    }
+    Ok(None)
+}
+
+pub(crate) fn extract_enum_from_subject(
+    asm_stmt: &AssembledStatement,
+    _scope: &Scope,
+) -> Result<Option<(AnalyzedExpr, GlossaType)>, GlossaError> {
+    if let Some(ref subj) = asm_stmt.subject
+        && let Some(result) = detect_enum_variant(subj, &asm_stmt.literals)
+    {
+        return Ok(Some(result));
+    }
+    Ok(None)
+}
+
+pub(crate) fn extract_genitive_method(
+    asm_stmt: &AssembledStatement,
+    scope: &Scope,
+) -> Result<Option<(AnalyzedExpr, GlossaType)>, GlossaError> {
+    if let Some(result) = try_parse_genitive_method_call(asm_stmt, scope) {
+        return Ok(Some(result));
+    }
+    Ok(None)
+}
+
+pub(crate) fn extract_enum_from_nominatives(
+    asm_stmt: &AssembledStatement,
+    _scope: &Scope,
+) -> Result<Option<(AnalyzedExpr, GlossaType)>, GlossaError> {
+    for nom in &asm_stmt.nominatives {
+        if let Some(result) = detect_enum_variant(nom, &asm_stmt.literals) {
+            return Ok(Some(result));
+        }
+    }
+    Ok(None)
+}
+
+pub(crate) fn extract_property_access(
+    asm_stmt: &AssembledStatement,
+    _scope: &Scope,
+) -> Result<Option<(AnalyzedExpr, GlossaType)>, GlossaError> {
+    let Some((owner, method)) = asm_stmt.property_accesses.first() else {
+        return Ok(None);
+    };
+
+    let receiver = AnalyzedExpr {
+        expr: AnalyzedExprKind::Variable(owner.clone().into()),
+        glossa_type: GlossaType::Unknown,
+    };
+    Ok(Some((
+        AnalyzedExpr {
+            expr: AnalyzedExprKind::MethodCall {
+                receiver: Box::new(receiver),
+                method: method.clone().into(),
+                args: vec![],
+            },
+            glossa_type: GlossaType::Number,
+        },
+        GlossaType::Number,
+    )))
+}
+
+pub(crate) fn extract_index_access(
+    asm_stmt: &AssembledStatement,
+    scope: &Scope,
+) -> Result<Option<(AnalyzedExpr, GlossaType)>, GlossaError> {
+    if let Some((array_expr, index_expr)) = asm_stmt.index_accesses.first() {
+        let array_analyzed = analyze_argument_expr(array_expr, scope)?;
+        let index_analyzed = analyze_argument_expr(index_expr, scope)?;
+        return Ok(Some((
+            AnalyzedExpr {
+                expr: AnalyzedExprKind::IndexAccess {
+                    array: Box::new(array_analyzed),
+                    index: Box::new(index_analyzed),
+                },
+                glossa_type: GlossaType::Unknown, // Element type is unknown without inference
+            },
+            GlossaType::Unknown,
+        )));
+    }
+    Ok(None)
+}
+
+pub(crate) fn extract_array(
+    asm_stmt: &AssembledStatement,
+    scope: &Scope,
+) -> Result<Option<(AnalyzedExpr, GlossaType)>, GlossaError> {
+    let Some(array_elements) = asm_stmt.arrays.first() else {
+        return Ok(None);
+    };
+
+    let analyzed_elements: Vec<AnalyzedExpr> = array_elements
+        .iter()
+        .map(|e| analyze_argument_expr(e, scope))
+        .collect::<Result<Vec<_>, _>>()?;
+
+    let element_type = analyzed_elements
+        .first()
+        .map(|e| e.glossa_type.clone())
+        .unwrap_or(GlossaType::Unknown);
+
+    Ok(Some((
+        AnalyzedExpr {
+            expr: AnalyzedExprKind::ArrayLiteral(analyzed_elements),
+            glossa_type: GlossaType::List(Box::new(element_type)),
+        },
+        GlossaType::List(Box::new(GlossaType::Unknown)),
+    )))
+}
+
+pub(crate) fn extract_binary_op(
+    asm_stmt: &AssembledStatement,
+    scope: &Scope,
+) -> Result<Option<(AnalyzedExpr, GlossaType)>, GlossaError> {
+    if asm_stmt.operators.is_empty() {
+        return Ok(None);
+    }
+
+    // Check if we can build from literals alone (2+ literals)
+    if asm_stmt.literals.len() >= 2 {
+        let exprs =
+            build_expressions_from_literals_and_ops(&asm_stmt.literals, &asm_stmt.operators)?;
+        if let Some(expr) = exprs.into_iter().next() {
+            let ty = expr.glossa_type.clone();
+            return Ok(Some((expr, ty)));
+        }
+    }
+
+    let make_var = |lemma: &smol_str::SmolStr| AnalyzedExpr {
+        expr: AnalyzedExprKind::Variable(lemma.clone()),
+        glossa_type: scope.lookup(lemma).cloned().unwrap_or(GlossaType::Unknown),
+    };
+
+    let op = asm_stmt.operators[0];
+
+    // Or check if we can combine object + literal with operator
+    if let Some(ref obj) = asm_stmt.object {
+        if !asm_stmt.literals.is_empty() {
+            // Build: object op literal
+            let left = make_var(&obj.lemma);
+            let right = literal_to_analyzed_expr(&asm_stmt.literals[0]);
+            let bin_expr = build_binary_expr(left, op, right);
+            let ty = bin_expr.glossa_type.clone();
+            return Ok(Some((bin_expr, ty)));
+        }
+
+        // Object + Nominative (e.g. x + y)
+        if let Some(nom) = asm_stmt.nominatives.first() {
+            let left = make_var(&obj.lemma);
+            let right = make_var(&nom.lemma);
+            let bin_expr = build_binary_expr(left, op, right);
+            let ty = bin_expr.glossa_type.clone();
+            return Ok(Some((bin_expr, ty)));
+        }
+    }
+
+    // Nominative + Nominative (e.g. a + b, where both are extra nominatives)
+    if asm_stmt.nominatives.len() >= 2 {
+        let left = make_var(&asm_stmt.nominatives[0].lemma);
+        let right = make_var(&asm_stmt.nominatives[1].lemma);
+        let bin_expr = build_binary_expr(left, op, right);
+        let ty = bin_expr.glossa_type.clone();
+        return Ok(Some((bin_expr, ty)));
+    }
+
+    Ok(None)
+}
+
+pub(crate) fn extract_literal(
+    asm_stmt: &AssembledStatement,
+    _scope: &Scope,
+) -> Result<Option<(AnalyzedExpr, GlossaType)>, GlossaError> {
+    if let Some(lit) = asm_stmt.literals.first() {
+        return Ok(Some((literal_to_analyzed_expr(lit), literal_to_type(lit))));
+    }
+    Ok(None)
+}
+
+pub(crate) fn extract_enum_from_object(
+    asm_stmt: &AssembledStatement,
+    _scope: &Scope,
+) -> Result<Option<(AnalyzedExpr, GlossaType)>, GlossaError> {
+    if let Some(ref obj) = asm_stmt.object {
+        return Ok(detect_enum_variant(obj, &asm_stmt.literals));
+    }
+    Ok(None)
+}
+
+pub(crate) fn extract_object_fallback(
+    asm_stmt: &AssembledStatement,
+    scope: &Scope,
+) -> Result<Option<(AnalyzedExpr, GlossaType)>, GlossaError> {
+    let Some(ref obj) = asm_stmt.object else {
+        return Ok(None);
+    };
+
+    let obj_lemma = &obj.lemma;
+
+    // Check if it's a numeral word
+    if let Some(value) = crate::morphology::lexicon::numeral_value(obj_lemma) {
+        return Ok(Some((
+            AnalyzedExpr {
+                expr: AnalyzedExprKind::NumberLiteral(value),
+                glossa_type: GlossaType::Number,
+            },
+            GlossaType::Number,
+        )));
+    }
+
+    if !scope.is_defined(obj_lemma) {
+        return Err(GlossaError::undefined(obj_lemma.as_str()));
+    }
+
+    Ok(Some((
+        AnalyzedExpr {
+            expr: AnalyzedExprKind::Variable(obj.lemma.clone()),
+            glossa_type: GlossaType::Unknown,
+        },
+        GlossaType::Unknown,
+    )))
+}
+
+/// Extract value from assembled statement
+///
+/// This function looks at the fields of the [`AssembledStatement`] and tries
+/// to extract a single meaningful value from it. It prioritizes different kinds
+/// of expressions in the following order:
+///
+/// 1. **Unwraps**: `expr!`
+/// 2. **Enum Variants**: `Some(val)`, `Ok(val)`, `None` (on subject or nominatives)
+/// 3. **Genitive Methods**: `owner.method`
+/// 4. **Property Access**: `user.name`
+/// 5. **Index Access**: `arr[0]`
+/// 6. **Array Literals**: `[1, 2]`
+/// 7. **Binary Operations**: `1 + 2`
+/// 8. **Object Enum Variants**: `Some(val)`, `Ok(val)`, `None` (on object) - *Prioritized over literals*
+/// 9. **Literals**: `42`, `"hello"`
+/// 10. **Variables (Object)**: `x`
+///
+/// Consolidates scattering values (numbers, strings, blocks) into a single logical expression.
+///
+/// In GLOSSA, depending on the sentence phrasing, the "value" of an assignment might be located
+/// in the subject slot, an explicit number literal slot, a string slot, or nested inside a phrase.
+/// This function acts as a semantic vacuum, pulling out the first valid expression value it can find
+/// in the statement regardless of where the `Assembler` categorized it grammatically.
+///
+/// # Returns
+///
+/// * `Ok((AnalyzedExpr, GlossaType))` containing the resolved expression and its inferred type.
+/// * `Err(GlossaError)` if no valid value expression can be identified.
+///
+/// # Examples
+///
+/// ```rust,ignore
+/// // Example cannot be run as a doctest because this module is pub(crate)
+/// use glossa::semantic::assembly::AssembledStatement;
+/// use glossa::semantic::conversion::extract_value;
+/// use glossa::semantic::resolver::Scope;
+/// use glossa::semantic::types::GlossaType;
+/// use glossa::semantic::AnalyzedExprKind;
+///
+/// let scope = Scope::new();
+/// let mut asm = AssembledStatement::new();
+///
+/// // Simulate a statement that contains a number literal: 42
+/// asm.numbers.push(42);
+///
+/// let (expr, ty) = extract_value(&asm, &scope).unwrap();
+///
+/// assert_eq!(ty, GlossaType::Number);
+/// match expr.expr {
+///     AnalyzedExprKind::NumberLiteral(n) => assert_eq!(n, 42),
+///     _ => panic!("Expected NumberLiteral"),
+/// }
+/// ```
+pub(crate) fn extract_value(
+    asm_stmt: &AssembledStatement,
+    scope: &Scope,
+) -> Result<(AnalyzedExpr, GlossaType), GlossaError> {
+    if !asm_stmt.nested_phrases.is_empty() {
+        // Handle nested phrases (parenthesized expressions) which act as values
+        // Usually there is only one for a value expression
+        if let Some(terms) = asm_stmt.nested_phrases.first() {
+            let phrase_expr = Expr::Phrase(terms.clone());
+            // Analyze with recursion depth check reset (as it's a new analysis root)
+            let analyzed = analyze_argument_expr(&phrase_expr, scope)?;
+            let ty = analyzed.glossa_type.clone();
+            return Ok((analyzed, ty));
+        }
+    }
+
+    if !asm_stmt.blocks.is_empty() {
+        // Handle blocks (braced expressions) which act as values
+        if let Some(stmts) = asm_stmt.blocks.first() {
+            let block_expr = Expr::Block(stmts.clone());
+            // Analyze with recursion depth check reset (as it's a new analysis root)
+            // Note: analyze_argument_expr will call analyze_block, which now enforces single-statement logic
+            let analyzed = analyze_argument_expr(&block_expr, scope)?;
+            let ty = analyzed.glossa_type.clone();
+            return Ok((analyzed, ty));
+        }
+    }
+
+    if let Some(res) = extract_unwrap(asm_stmt, scope)? {
+        return Ok(res);
+    }
+    if let Some(res) = extract_enum_from_subject(asm_stmt, scope)? {
+        return Ok(res);
+    }
+    if let Some(res) = extract_genitive_method(asm_stmt, scope)? {
+        return Ok(res);
+    }
+    if let Some(res) = extract_enum_from_nominatives(asm_stmt, scope)? {
+        return Ok(res);
+    }
+    if let Some(res) = extract_property_access(asm_stmt, scope)? {
+        return Ok(res);
+    }
+    if let Some(res) = extract_index_access(asm_stmt, scope)? {
+        return Ok(res);
+    }
+    if let Some(res) = extract_array(asm_stmt, scope)? {
+        return Ok(res);
+    }
+    if let Some(res) = extract_binary_op(asm_stmt, scope)? {
+        return Ok(res);
+    }
+    // Fix: Check object for enum variants BEFORE literals to avoid shadowing Some(literal) by literal
+    if let Some(res) = extract_enum_from_object(asm_stmt, scope)? {
+        return Ok(res);
+    }
+    if let Some(res) = extract_literal(asm_stmt, scope)? {
+        return Ok(res);
+    }
+    if let Some(res) = extract_object_fallback(asm_stmt, scope)? {
+        return Ok(res);
+    }
+
+    // Default
+    Ok((
+        AnalyzedExpr {
+            expr: AnalyzedExprKind::NumberLiteral(0),
+            glossa_type: GlossaType::Number,
+        },
+        GlossaType::Number,
+    ))
+}
